Allow moving examples directory
Use a new environment variable OPENCMISSEXAMPLES_ROOT to define the path to
the OpenCMISS examples in anticipation of moving them to a separate repository.
For now this will default to $OPENCMISS_ROOT/cm/examples.

Also allow example Makefiles to overwrite EXAMPLE_ROOT, so that
people can use the ExampleMakefile for programs outside of the examples
directory.



git-svn-id: https://opencmiss.svn.sourceforge.net/svnroot/opencmiss/cm/trunk@2175 61e36070-c886-4827-b7c0-ed9e4dc59cc8