Change equations_set%boundary conditions to solver_equations%boundary_conditions
The only place where this is still required is in finite elasticity
residual evaluate. Once this is fixed the pointer from an equations
set to the boundary conditions can be removed.

Had to remove the equations set boundary conditions analytic calculation
out of the equations set setup routine and into it's own routine so that
the boundary conditions can be passed in.