Python: Initialise null CMISS type if no args passed to creator
This is to allow passing null CMISS types as arguments to routines
like CMISSProblemSolverGet, which returns the solver inside the
buffer CMISS type. Some types also have a CreateStart routine that
can be called so we can't just check if there's a CreateStart and
try to call it. Tracker item 2147.