Code to attach a coordinate system to an interface type
Code has been added to define and retrieve a coordinate system in an interface region.
"generated_mesh_routines.f90" has also been changed since the parameters of the generated mesh needs to be consistent with those of the CS.

tracker item: 3118

Could I ask Prasad to review the code please?