Updating to allow for coupling of dependent field variables whos number of components do not match
For example in problems such as 1)incompressible finite elasticity - 2) Membrane coupling where 1) has an extra dependent field variable component relating to hydrostatic pressure
In general the user will need to set a fixed boundary condition on the interface Lagrange field dof relating to the field components not present in each of the coupled bodies.
This would eliminate the dof from the solver matrices, ie in the example above the hydrostatic pressure term from 1) would be removed.

Tracker item 2271