Reimplement Neumann point based conditions
The previous implementation did not integrate with other boundary
condition types and was difficult to understand to fix, so the
Neumann point conditions have been re-written from scratch.

This adds a new BoundaryConditions_NeumannIntegrate routine
that integrates point Neumann conditions over faces in 3D and
lines in 2D, and a BoundaryConditionsNeumannType for storing
information about the Neumann point conditions.

Tracker items 1809 and 2111.