Correct sign of right hand side field values in solver RHS
All equations sets should be standardised on using f(x) - b = 0 for
nonlinear problems, or Kx = b for linear problems. So the right hand
side vector values should be directly added to the solver right hand
side without multiplying by negative one.

This corrects the direction of forces for elasticity problems so that
tensile forces are positive, so some programs might break and require
the sign of Neumann conditions to be changed.