Performance improvement when using diagnostics or timing
With diagnostics or timing enabled, about 70% of the run time was spent
in the iso_varying_string module comparing the name of the diagnostic or
timing routine to the current routine in the enters routine. This took a
very long time as the strings were converted to char arrays for
comparison, which required allocating memory.

This commit does the comparison without converting to a char array and
gives a large speed improvement when using diagnostics or timing. An
example with two routines timed showed a 3 times overall speedup.

Now most of the time spent in enters (57%) is spent creating a varying
string from a character array (op_assign_vs_ch).