Use NumPy arrays in the Python bindings
Replace the SWIG array typemaps with ones from NumPy's numpy.i file.
This has been modified to support output of 2D Fortran arrays and to use
a tuple for output of multiple parameters.
Input arrays are also now declared as const.
In order to work with the the NumPy typemaps, the C bindings have been
changed so that arrays of a known size, eg. see
"CMISSAnalyticAnalysisRmsErrorGetNode" for example use the square
bracket array syntax rather than pointer syntax.

This change requires NumPy when using the OpenCMISS Python bindings and
the NumPy header files in order to compile the Python bindings.

This will allow us to provide a "view" on data that has been allocated
within the library, eg. CMISSField_ParameterSetDataGet can be supported,
although this hasn't been implemented yet.

Tracker item 2147.