Add redirection for Python field routines
Allow calling the generic interface method when there are different
methods for different data types, eg. Field.ComponentValuesInitialise
can be called and will check the field variable data type, then call
ComponentValuesInitialiseDP or ComponentValuesInitialiseIntg etc.

Unfortunately I don't think it's possible to automate this in a general
way as we want to choose the routine based on the actual field data
type, not the type of input parameter. For example, if we have a field
with DP data type and we call the initialise routine with an integer
from Python, we still want it to call the InitialiseDP routine and cast
the Python integer to a double.

This means that this file will have to be manually updated as required.

Tracker item 2147.