Added MESH_TOPOLOGY_ELEMENTS_ADJACENT_ELEMENT_GET routine and necessary library routines
Currently returns only the first adjacent element and needs to be extended to output all or the selected number of adjacent elements.
An additional routine can be added to first output the number of adjacent elements. This can subsequently be used to define array sizes required for the adjacent element get routine.
This is useful/required in some circumstances eg the python bindings require the output array size of the adjacent element array to be known a-prior as these are arguments to the OpenCMISS python library calls.

Tracker Item 3427