Revert "Fix crash when loading Intel MKL lib in Python"
Tracker item 3504.

Using the RTLD_GLOBAL option when loading dynamic libraries causes a
segfault in SciPy if SciPy is loaded after OpenCMISS.
Instead of loading libraries with RTLD_GLOBAL, we can instead link to
the single dynamic libmkl_rt to avoid problems loading Intel MKL
dynamically with Python.

There's still a crash when calling SciPyRestore on a DistributedMatrix
so this might be due to a change with the new libraries, rather than
anything related to this RTLD_GLOBAL issue.

This reverts commit 401637b88c695edc4d946119bd299b33747e8055.